Last update

+++ app/Http/Controllers/Api/DirectoryController.php
@@ -90,16 +90,21 @@ class DirectoryController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Directory $directory)
+    public function destroy(Request  $request, Directory $directory)
     {
-        $response = Gate::inspect('destroy', $directory);
+        $response = Gate::inspect('update', $directory);
 
         if (!$response->allowed()) {
             return response()->json(['error' => __('This action is unauthorized')], 403);
         }
 
+        $user_license = $request->user()->user_license;
+
+        if (!$user_license)
+            return response()->json(['error' => __('No posee licencia')], 403);
+
         $directory->delete();
 
-        return new DirectoryResource($directory);
+        return DirectoryResource::collection($user_license->directories()->paginate(25));
     }
 }
